+++ packages/amplication-git-utils/src/git/dto/create-pr.dto.ts
@@ -1,0 +1,6 @@
+export class CreatePrDto {
+  title: string;
+  body: string;
+  head: string;
+  base?: string | undefined;
+}
