+++ packages/amplication-git-utils/src/git/git.service.ts
@@ -206,54 +206,4 @@ export class GitService {
     const service = this.gitServiceFactory.getService(gitProvider);
     return service.getRepository(installationId, owner, repo);
   }
-
-  createBranch(
-    gitProvider: EnumGitProvider,
-    installationId: string,
-    owner: string,
-    repo: string,
-    newBranchName: string,
-    sha?: string
-  ): Promise<Branch> {
-    const service = this.gitServiceFactory.getService(gitProvider);
-    return service.createBranch(
-      installationId,
-      owner,
-      repo,
-      newBranchName,
-      sha
-    );
-  }
-
-  isBranchExist(
-    gitProvider: EnumGitProvider,
-    installationId: string,
-    owner: string,
-    repo: string,
-    branch: string
-  ) {
-    const service = this.gitServiceFactory.getService(gitProvider);
-    return service.isBranchExist(installationId, owner, repo, branch);
-  }
-
-  getBranch(
-    gitProvider: EnumGitProvider,
-    installationId: string,
-    owner: string,
-    repo: string,
-    branch: string
-  ): Promise<Branch> {
-    const service = this.gitServiceFactory.getService(gitProvider);
-    return service.getBranch(installationId, owner, repo, branch);
-  }
-
-  getFirstDefaultBranchCommit(
-    gitProvider: EnumGitProvider,
-    installationId: string,
-    owner: string,
-    repo: string
-  ) {
-    const service = this.gitServiceFactory.getService(gitProvider);
-    return service.getFirstDefaultBranchCommit(installationId, owner, repo);
-  }
 }
