+++ packages/amplication-git-utils/src/git/github.service.ts
@@ -303,173 +303,4 @@ export class GithubService {
       defaultBranch,
     };
   }
-
-  async createBranch(
-    installationId: string,
-    owner: string,
-    repo: string,
-    newBranchName: string,
-    sha?: string
-  ): Promise<Branch> {
-    const octokit = await this.getInstallationOctokit(installationId);
-    let baseSha = sha;
-    if (!baseSha) {
-      const repository = await this.getRepository(installationId, owner, repo);
-      const { defaultBranch } = repository;
-
-      const refs = await octokit.rest.git.getRef({
-        owner,
-        repo,
-        ref: `heads/${defaultBranch}`,
-      });
-      baseSha = refs.data.object.sha;
-    }
-    const { data: branch } = await octokit.rest.git.createRef({
-      owner,
-      repo,
-      ref: `refs/heads/${newBranchName}`,
-      sha: baseSha,
-    });
-    return { name: newBranchName, sha: branch.object.sha };
-  }
-
-  async isBranchExist(
-    installationId: string,
-    owner: string,
-    repo: string,
-    branch: string
-  ): Promise<boolean> {
-    try {
-      const refs = await this.getBranch(installationId, owner, repo, branch);
-      return Boolean(refs);
-    } catch (error) {
-      return false;
-    }
-  }
-
-  async getBranch(
-    installationId: string,
-    owner: string,
-    repo: string,
-    branch: string
-  ): Promise<Branch> {
-    const octokit = await this.getInstallationOctokit(installationId);
-    const refs = await octokit.rest.git.getRef({
-      owner,
-      repo,
-      ref: `heads/${branch}`,
-    });
-    return { sha: refs.data.object.sha, name: branch };
-  }
-
-  async getFirstDefaultBranchCommit(
-    installationId: string,
-    owner: string,
-    repo: string
-  ): Promise<{ sha: string }> {
-    const octokit = await this.getInstallationOctokit(installationId);
-    const { defaultBranch } = await this.getRepository(
-      installationId,
-      owner,
-      repo
-    );
-    const firstCommit: TData = await octokit.graphql(
-      `query ($owner: String!, $repo: String!, $defaultBranch: String!) {
-      repository(name: $repo, owner: $owner) {
-        ref(qualifiedName: $defaultBranch) {
-          target {
-            ... on Commit {
-              history(first: 1) {
-                nodes {
-                  oid
-                  url
-                }
-                totalCount
-                pageInfo {
-                  endCursor
-                }
-              }
-            }
-          }
-        }
-      }
-    }`,
-      {
-        owner,
-        repo,
-        defaultBranch,
-      }
-    );
-    const {
-      repository: {
-        ref: {
-          target: {
-            history: { nodes: firstCommitNodes, pageInfo, totalCount },
-          },
-        },
-      },
-    } = firstCommit;
-
-    if (totalCount <= 1) {
-      return { sha: firstCommitNodes[0].oid };
-    }
-
-    const cursorPrefix = pageInfo.endCursor.split(" ")[0];
-    const nextCursor = `${cursorPrefix} ${totalCount - 2}`;
-
-    const lastCommitData: TData = await octokit.graphql(
-      `query ($owner: String!, $repo: String!, $defaultBranch: String!, $nextCursor: String!) {
-        repository(name: $repo, owner: $owner) {
-          ref(qualifiedName: $defaultBranch) {
-            target {
-              ... on Commit {
-                history(first: 1, after: $nextCursor) {
-                  nodes {
-                    oid
-                    url
-                  }
-                  totalCount
-                  pageInfo {
-                    endCursor
-                  }
-                }
-              }
-            }
-          }
-        }
-      }`,
-      { owner, repo, defaultBranch, nextCursor }
-    );
-    const {
-      repository: {
-        ref: {
-          target: {
-            history: { nodes: lastCommitNodes },
-          },
-        },
-      },
-    } = lastCommitData;
-    return { sha: lastCommitNodes[0].oid };
-  }
 }
-
-type TData = {
-  repository: {
-    ref: {
-      target: {
-        history: {
-          nodes: [
-            {
-              oid: string;
-              url: string;
-            }
-          ];
-          pageInfo: {
-            endCursor: string;
-          };
-          totalCount: number;
-        };
-      };
-    };
-  };
-};
